i3c: master: cdns: Fix potential sign extension in THD_DELAY logic
The cdns_i3c_master_calculate_thd_delay() function returns
a u8. Due to C's integer promotion rules, this value is
promoted to a signed 'int' before being passed to the
CTRL_THD_DELAY() macro.

If the value is large, this can lead to sign-extension
when the value is bit-shifted and OR'd into the u32 'ctrl'
register, corrupting other control bits.

Fix this by explicitly casting the return value to (u32)
before the macro is applied, ensuring no sign extension
can occur.

+++ drivers/i3c/master/i3c-master-cdns.c
@@ -1303,8 +1303,11 @@ static int cdns_i3c_master_bus_init(struct i3c_master_controller *m)
 	 * MIPI I3C Specification 1.0 defines non-zero minimal tHD_PP timing on
 	 * master output. This setting allows to meet this timing on master's
 	 * SoC outputs, regardless of PCB balancing.
+	 *
+	 * Fix: Cast to unsigned int to prevent sign extension during bit shift
 	 */
-	ctrl |= CTRL_THD_DELAY(cdns_i3c_master_calculate_thd_delay(master));
+	ctrl |= CTRL_THD_DELAY(
+		(u32)cdns_i3c_master_calculate_thd_delay(master));
 	writel(ctrl, master->regs + CTRL);
 
 	cdns_i3c_master_enable(master);
